Changes performed, ready to be merged
