+++ app/models/Certification.js
@@ -1,0 +1,25 @@
+const mongoose = require('mongoose');
+const CONSTANTS = require('../../config/constants');
+
+const Schema = mongoose.Schema;
+const certificationSchema = new Schema(
+    {
+        name: {
+            type: String,
+        },
+        authority: {
+            type: String,
+        },
+        lic_number: {
+            type: String,
+        },
+        url: {
+            type: String,
+        },
+    },
+    {
+        _id: false,
+    },
+);
+
+module.exports = certificationSchema;
